+++ src/typeparser/structs.ts
@@ -7,7 +7,7 @@ export class StructureParser {
     const output: structureType[] = [];
     const allLines = docText.split("\n");
     for (const line of allLines) {
-      const regex = /#Struct\s*([\w:]+)\s*(as\s+(\w+)\s*){0,1}(\{){0,1}/;      
+      const regex = /#Struct\s*([\w:]+)\s*(as\s+(\w+)\s*){0,1}(\{){0,1}/;
       const m = regex.exec(line);
       let name = "";
       if (m != null) {
@@ -54,55 +54,12 @@ export class StructureParser {
           range: range,
           structName: name,
           members: elemTypes,
-          codeBlock: m[0] + structInfo +"\n}",
+          codeBlock: m[0] + structInfo + "\n}",
           ext: isExt,
           extType: extType,
         });
       }
     }
-    /*  const structLines = docText.match(
-      /#Struct\s*\b(.*?)\b\s*(as\s+(\w+)\s*){0,1}[{]/gm
-    );
-    if (structLines !== null) {
-      structLines.forEach((struct) => {
-        const name = struct.split(/\s/)[1];
-        const start = docText.indexOf(struct);
-        let end = start;
-        for (end; end < docText.length && docText.charAt(end) != "}"; end++);
-
-        if (docText.charAt(end) == "}") {
-          const structInfo: string = docText.slice(start + struct.length, end);
-          const variableRegex = /(\w+(?:\[(?:\w+?)?\]){0,4})\s*(\w+)\s*;/gm;
-
-          variableRegex.lastIndex = 0;
-          const structVar = structInfo.match(variableRegex);
-          const elemTypes: nameType[] = [];
-          if (structVar !== null) {
-            structVar.forEach((vari) => {
-              variableRegex.lastIndex = 0;
-              const types = variableRegex.exec(vari);
-              if (types !== null) {
-                elemTypes.push({ name: types[2], type: types[1] });
-              }
-            });
-          }
-
-          const offset = docText.indexOf(struct.trim());
-          const lines = docText.slice(0, offset).split("\n");
-          const allLines = docText.split("\n");
-          const i = lines.length - 1;
-          const start2 = allLines[i].indexOf(name);
-          const range = new vscode.Range(i, start2, i, start2 + name.length);
-
-          output.push({
-            range: range,
-            structName: name,
-            members: elemTypes,
-            codeBlock: `#Struct ${name} {\n` + structInfo + "}\n",
-          });
-        }
-      });
-    } */
 
     return removeDuplicates(output, "structName");
   }